+++ PrivateFrameworks/HomeSharing.framework/HSIncrementRequest.h
@@ -5,8 +5,8 @@
 @interface HSIncrementRequest : HSRequest {
 }
 
-+ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2;
++ (id)requestWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2;
 
-- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned int)arg2;
+- (id)initWithDatabaseID:(unsigned int)arg1 itemID:(unsigned long long)arg2;
 
 @end
