+++ PrivateFrameworks/OfficeImport.framework/CPEnumerationMap.h
@@ -12,6 +12,6 @@
 - (id)initWithStructs:(const struct CPEnumerationStruct { int x1; id x2; }*)arg1 count:(int)arg2 caseSensitive:(BOOL)arg3;
 - (id)initWithStructs:(const struct CPEnumerationStruct { int x1; id x2; }*)arg1 count:(int)arg2;
 - (id)stringForValue:(int)arg1;
-- (int)valueForString:(id)arg1;
+- (long)valueForString:(id)arg1;
 
 @end
