+++ PrivateFrameworks/OfficeImport.framework/EDWorksheet.h
@@ -13,8 +13,8 @@
     EDCollection *mHyperlinks;
     EDReference *mImplicitCellArea;
     EDReference *mMaxCellReferencedInFormulas;
-    unsigned int mMaxColumnOutlineLevel;
-    unsigned int mMaxRowOutlineLevel;
+    unsigned long mMaxColumnOutlineLevel;
+    unsigned long mMaxRowOutlineLevel;
     EDMergedCellCollection *mMergedCells;
     TSUPointerKeyDictionary *mMergedCols;
     TSUPointerKeyDictionary *mMergedRows;
@@ -38,8 +38,8 @@
 - (BOOL)isColMerged:(unsigned int)arg1;
 - (BOOL)isRowMerged:(unsigned int)arg1;
 - (id)maxCellReferencedInFormulas;
-- (unsigned int)maxColumnOutlineLevel;
-- (unsigned int)maxRowOutlineLevel;
+- (unsigned long)maxColumnOutlineLevel;
+- (unsigned long)maxRowOutlineLevel;
 - (id)mergedCells;
 - (id)mergedColRef:(unsigned int)arg1;
 - (id)mergedRowRef:(unsigned int)arg1;
@@ -51,16 +51,16 @@
 - (void)setDefaultRowHeight:(unsigned short)arg1;
 - (void)setFitToPage:(bool)arg1;
 - (void)setImplicitCellArea:(id)arg1;
-- (void)setMaxColumnOutlineLevel:(unsigned int)arg1;
-- (void)setMaxRowOutlineLevel:(unsigned int)arg1;
+- (void)setMaxColumnOutlineLevel:(unsigned long)arg1;
+- (void)setMaxRowOutlineLevel:(unsigned long)arg1;
 - (void)setMergedColsRef:(id)arg1 from:(unsigned int)arg2 to:(unsigned int)arg3;
 - (void)setMergedRowsRef:(id)arg1 from:(unsigned int)arg2 to:(unsigned int)arg3;
 - (void)setPane:(id)arg1;
 - (void)setup;
 - (id)tables;
 - (void)teardown;
-- (void)updateMaxColumnOutlineLevelIfNeeded:(unsigned int)arg1;
-- (void)updateMaxRowOutlineLevelIfNeeded:(unsigned int)arg1;
+- (void)updateMaxColumnOutlineLevelIfNeeded:(unsigned long)arg1;
+- (void)updateMaxRowOutlineLevelIfNeeded:(unsigned long)arg1;
 - (id)worksheetWarnings;
 
 @end
